requirements.txt
djangorestframework
drf-spectacular
psycopg2
ruff
pytest
pytest-django
django-extensions  # For shell_plus
djangorestframework-simplejwt